+++ RELEASE_NOTES.rst
@@ -80,6 +80,9 @@ Additions and enhancements to dendrogram and matrix visualizations.
   private method ``_get_style_kwargs()``
 - API: implement :class:`.RgbColor` and :class:`.RgbaColor` as classes instead of
   type aliases
+- API: remove method ``dark()`` from class :class:`.ColoredStyle` and instead introduce
+  constants :attr:`.ColorScheme.DEFAULT`, :attr:`.ColorScheme.DEFAULT_LIGHT`, and
+  :attr:`.ColorScheme.DEFAULT_DARK`
 
 
 *pytools* 1.2